+++ src/http2utils.h
@@ -8,5 +8,5 @@
 
 int verify_setting(uint16_t id, uint32_t value);
 int read_n_bytes(uint8_t *buff_read, int n);
-int http_write(uint8_t *bytes, uint8_t length);//here just for the code to compile
-int http_read(uint8_t *bytes, uint8_t length);//here just for the code to compile
+int http_write(uint8_t *bytes, int length);//here just for the code to compile
+int http_read(uint8_t *bytes, int length);//here just for the code to compile
